PWEB-5408 - cleanup ios8 warnings for the samples

+++ DDx/DDx/DDxRoundtripPing.h
@@ -8,7 +8,7 @@
 
 #import <Foundation/Foundation.h>
 
-typedef void (^DDxPingCompletion)(int numPings, double average, NSArray *times);
+typedef void (^DDxPingCompletion)(NSInteger numPings, double average, NSArray *times);
 
 @interface DDxRoundtripPing : NSObject
